+++ surfh/models.py
@@ -32,8 +32,7 @@ from . import cython_2D_interpolation
 from . import cythons_files
 from . import shared_dict
 
-#from .CFAsyncProcessPool import CFAPP
-#from .AsyncProcessPool import APP
+from .AsyncProcessPoolLight import APPL
 from multiprocessing import Pool
 from multiprocessing import Process, Queue, connection
 
@@ -246,6 +245,7 @@ class Channel(LinOp):
         wavel_axis: array,
         srf: int,
         pointings: instru.CoordList,
+        shared_metadata_path: str,
     ):
         """Forward model of a Channel
 
@@ -269,9 +269,19 @@ class Channel(LinOp):
         alpha and beta axis must have the same step and must be regular. This is
         not the case for wavel_axis that must only have incrising values.
         """
-        self.wavel_axis = wavel_axis
-        self.alpha_axis = alpha_axis
-        self.beta_axis = beta_axis
+
+
+        _metadata = shared_dict.attach(shared_metadata_path)
+        self._metadata_path = _metadata.path
+        _metadata["wavel_axis"] = wavel_axis
+        _metadata["alpha_axis"] = alpha_axis
+        _metadata["beta_axis"] = beta_axis
+
+        #self.wavel_axis = wavel_axis
+        #self.alpha_axis = alpha_axis
+        #self.beta_axis = beta_axis
+
+        
 
         if alpha_axis[1] - alpha_axis[0] != beta_axis[1] - beta_axis[0]:
             logger.warning(
@@ -283,9 +293,11 @@ class Channel(LinOp):
 
         self.srf = srf
         self.imshape = (len(alpha_axis), len(beta_axis))
-        self._otf_sr = udft.ir2fr(np.ones((srf, 1)), self.imshape)[np.newaxis, ...]
 
-        self.local_alpha_axis, self.local_beta_axis = self.instr.fov.local_coords(
+        _metadata["_otf_sr"] = udft.ir2fr(np.ones((srf, 1)), self.imshape)[np.newaxis, ...]
+        #self._otf_sr = udft.ir2fr(np.ones((srf, 1)), self.imshape)[np.newaxis, ...]
+
+        _metadata["local_alpha_axis"], _metadata["local_beta_axis"] = self.instr.fov.local_coords(
             self.step,
             alpha_margin=5 * self.step,
             beta_margin=5 * self.step,
@@ -306,24 +318,24 @@ class Channel(LinOp):
         self.local_shape = (
             # self.instr.n_wavel,
             self.wslice.stop - self.wslice.start,
-            len(self.local_alpha_axis),
-            len(self.local_beta_axis),
+            len(_metadata["local_alpha_axis"]),
+            len(_metadata["local_beta_axis"]),
         )
 
-        _data = shared_dict.create("data")
-        _data["test"] = np.zeros(oshape)
-        self._data_path = _data.path
-
-
+        _metadata["fw_data"] = np.zeros(oshape)
+        _metadata["ad_data"] = np.zeros(ishape, dtype=np.complex128)
+ 
         super().__init__(ishape, oshape, self.instr.name)
 
     @property
     def step(self) -> float:
-        return self.alpha_axis[1] - self.alpha_axis[0]
+        alpha_axis = shared_dict.attach(self._metadata_path)["alpha_axis"]
+        return alpha_axis[1] - alpha_axis[0]
 
     @property
     def beta_step(self) -> float:
-        return self.beta_axis[1] - self.beta_axis[0]
+        beta_axis = shared_dict.attach(self._metadata_path)["beta_axis"]
+        return beta_axis[1] - beta_axis[0]
 
     @property
     def n_alpha(self) -> int:
@@ -333,13 +345,15 @@ class Channel(LinOp):
     @property
     def wslice(self) -> slice:
         """The wavelength slice of input that match instr with 0.1 μm of margin."""
-        return self.instr.wslice(self.wavel_axis, 0.1)
+        wavel_axis = shared_dict.attach(self._metadata_path)["wavel_axis"]
+        return self.instr.wslice(wavel_axis, 0.1)
 
     @property
     def npix_slit(self) -> int:
         """The number of pixel inside a slit"""
+        beta_axis = shared_dict.attach(self._metadata_path)["beta_axis"]
         return int(
-            ceil(self.instr.slit_beta_width / (self.beta_axis[1] - self.beta_axis[0]))
+            ceil(self.instr.slit_beta_width / (beta_axis[1] - beta_axis[0]))
         )
 
     def slit_local_fov(self, slit_idx) -> instru.LocalFOV:
@@ -349,16 +363,18 @@ class Channel(LinOp):
 
     def slit_slices(self, slit_idx: int) -> Tuple[slice, slice]:
         """The slices of slit `slit_idx` in local axis"""
+        local_alpha_axis = shared_dict.attach(self._metadata_path)["local_alpha_axis"]
+        local_beta_axis = shared_dict.attach(self._metadata_path)["local_beta_axis"]
         slices = self.slit_local_fov(slit_idx).to_slices(
-            self.local_alpha_axis, self.local_beta_axis
+            local_alpha_axis, local_beta_axis
         )
         # If slice to long, remove one pixel at the beginning or the end
         if (slices[1].stop - slices[1].start) > self.npix_slit:
             if abs(
-                self.local_beta_axis[slices[1].stop]
+                local_beta_axis[slices[1].stop]
                 - self.slit_local_fov(slit_idx).beta_end
             ) > abs(
-                self.local_beta_axis[slices[1].start]
+                local_beta_axis[slices[1].start]
                 - self.slit_local_fov(slit_idx).beta_start
             ):
                 slices = (slices[0], slice(slices[1].start, slices[1].stop - 1))
@@ -377,13 +393,15 @@ class Channel(LinOp):
 
     def slit_weights(self, slit_idx: int) -> array:
         """The weights of slit `slit_idx` in local axis"""
+        local_alpha_axis = shared_dict.attach(self._metadata_path)["local_alpha_axis"]
+        local_beta_axis = shared_dict.attach(self._metadata_path)["local_beta_axis"]
         slices = self.slit_slices(slit_idx)
 
         weights = fov_weight(
             self.slit_local_fov(slit_idx),
             slices,
-            self.local_alpha_axis,
-            self.local_beta_axis,
+            local_alpha_axis,
+            local_beta_axis,
         )
 
         # If previous do not share a pixel
@@ -424,8 +442,13 @@ class Channel(LinOp):
     def gridding(self, inarray: array, pointing: instru.Coord) -> array:
         """Returns interpolation of inarray in local referential"""
         # α and β inside the FOV shifted to pointing, in the global ref.
+        local_alpha_axis = shared_dict.attach(self._metadata_path)["local_alpha_axis"]
+        local_beta_axis = shared_dict.attach(self._metadata_path)["local_beta_axis"]
+        alpha_axis = shared_dict.attach(self._metadata_path)["alpha_axis"]
+        beta_axis = shared_dict.attach(self._metadata_path)["beta_axis"]
+
         alpha_coord, beta_coord = (self.instr.fov + pointing).local2global(
-            self.local_alpha_axis, self.local_beta_axis
+            local_alpha_axis, local_beta_axis
         )
 
         # Necessary for interpn to process 3D array. No interpolation is done
@@ -441,7 +464,7 @@ class Channel(LinOp):
             ]
         ).T          
 
-        return cython_2D_interpolation.interpn( (self.alpha_axis, self.beta_axis), 
+        return cython_2D_interpolation.interpn( (alpha_axis, beta_axis), 
                                               inarray, 
                                               local_coords, 
                                               len(wl_idx)).reshape(out_shape) 
@@ -450,15 +473,20 @@ class Channel(LinOp):
     def gridding_t(self, inarray: array, pointing: instru.Coord) -> array:
         """Returns interpolation of inarray in global referential"""
         # α and β inside the FOV shifted to pointing, in the global ref.
+        local_alpha_axis = shared_dict.attach(self._metadata_path)["local_alpha_axis"]
+        local_beta_axis = shared_dict.attach(self._metadata_path)["local_beta_axis"]
+        alpha_axis = shared_dict.attach(self._metadata_path)["alpha_axis"]
+        beta_axis = shared_dict.attach(self._metadata_path)["beta_axis"]
+
         alpha_coord, beta_coord = (self.instr.fov + pointing).global2local(
-            self.alpha_axis, self.beta_axis
+            alpha_axis, beta_axis
         )
 
         # Necessary for interpn to process 3D array. No interpolation is done
         # along that axis.
         wl_idx = np.arange(inarray.shape[0])
 
-        out_shape = (len(wl_idx), len(self.alpha_axis), len(self.beta_axis))
+        out_shape = (len(wl_idx), len(alpha_axis), len(beta_axis))
 
         global_coords = np.vstack(
             [
@@ -467,7 +495,7 @@ class Channel(LinOp):
             ]
         ).T
 
-        return cython_2D_interpolation.interpn( (self.local_alpha_axis, self.local_beta_axis), 
+        return cython_2D_interpolation.interpn( (local_alpha_axis, local_beta_axis), 
                                               inarray, 
                                               global_coords, 
                                               len(wl_idx),
@@ -477,23 +505,27 @@ class Channel(LinOp):
 
     def sblur(self, inarray_f: array) -> array:
         """Return spatial blurring of inarray_f in Fourier space for SR"""
+        _otf_sr = shared_dict.attach(self._metadata_path)["_otf_sr"]
         return idft(
-            inarray_f * self._otf_sr,
+            inarray_f * _otf_sr,
             self.imshape,
         )
 
 
     def sblur_t(self, inarray: array) -> array:
         """Return spatial blurring transpose of inarray for SR. Returns in Fourier space"""
-        return dft(inarray) * self._otf_sr.conj()
+        _otf_sr = shared_dict.attach(self._metadata_path)["_otf_sr"]
+        return dft(inarray) * _otf_sr.conj()
 
     def _wpsf(self, length: int, step: float) -> array:
         """Return spectral PSF"""
         # ∈ [0, β_s]
+        wavel_axis = shared_dict.attach(self._metadata_path)["wavel_axis"]
         beta_in_slit = np.arange(0, length) * step
+        
         return self.instr.spectral_psf(
             beta_in_slit - np.mean(beta_in_slit),  # ∈ [-β_s / 2, β_s / 2]
-            self.wavel_axis[self.wslice],
+            wavel_axis[self.wslice],
             arcsec2micron=self.instr.wavel_step / self.instr.det_pix_size,
         )
 
@@ -505,100 +537,18 @@ class Channel(LinOp):
         """Returns spectral blurring transpose of inarray"""
         return wblur_t(inarray, self._wpsf(inarray.shape[2], self.beta_step))
 
-
-    def forward(self, inarray_f: array):
-        """inarray is supposed in global coordinate, spatially blurred and in Fourier space.
-
-        Output is an array of shape (pointing, slit, wavelength, alpha)."""
-        # [pointing, slit, λ', α]
-        out = np.empty(self.oshape)
-        blurred = self.sblur(inarray_f[self.wslice, ...])
-        #logger.info(f"{self.name} : IDFT2({inarray_f.shape})")
-        for p_idx, pointing in enumerate(self.pointings):
-            #logger.info(
-            #    f"{self.name} : gridding [{p_idx}/{len(self.pointings)}] {blurred.shape} -> {(blurred.shape[0],) + self.local_shape[1:]}"
-            #)
-            gridded = self.gridding(blurred, pointing)
-            # Slit parallelization
-            for slit_idx in range(self.instr.n_slit):
-                # Slicing, weighting and α subsampling for SR
-                #logger.info(f"{self.name} : slicing [{slit_idx+1}/{self.instr.n_slit}]")
-                sliced = self.slicing(gridded, slit_idx)[
-                    :, : self.oshape[3] * self.srf : self.srf
-                ]
-                
-                out[p_idx, slit_idx, :, :] = self.instr.pce[
-                    ..., np.newaxis
-                ]* self.wblur(sliced).sum(axis=2)
-                                
-        return out
-
-
-    def forward_multiproc(self, inarray_f, idx, q):
-        """inarray is supposed in global coordinate, spatially blurred and in Fourier space.
-
-        Output is an array of shape (pointing, slit, wavelength, alpha)."""
-        # [pointing, slit, λ', α]
-        out = np.empty(self.oshape)
-        print("Idx = ", idx)
-        blurred = self.sblur(inarray_f[self.wslice, ...])
-        for p_idx, pointing in enumerate(self.pointings):
-            gridded = self.gridding(blurred, pointing)
-            for slit_idx in range(self.instr.n_slit):
-                sliced = self.slicing(gridded, slit_idx)[
-                    :, : self.oshape[3] * self.srf : self.srf
-                ]
-                
-                out[p_idx, slit_idx, :, :] = self.instr.pce[
-                    ..., np.newaxis
-                ]* self.wblur(sliced).sum(axis=2)
-
-           
-        #q.put({"idx": idx, "out" : out})
-        #q.put(out)
-        print("Return")   
-        return
-
-
-    def forward_parallel(self, inarray_f: array, idx: int) -> array:
-        """inarray is supposed in global coordinate, spatially blurred and in Fourier space.
-
-        Output is an array of shape (pointing, slit, wavelength, alpha)."""
-        # [pointing, slit, λ', α]
-        out = np.empty(self.oshape)
-        print("Idx = ", idx)
-        blurred = self.sblur(inarray_f[self.wslice, ...])
-        for p_idx, pointing in enumerate(self.pointings):
-            gridded = self.gridding(blurred, pointing)
-            for slit_idx in range(self.instr.n_slit):
-                sliced = self.slicing(gridded, slit_idx)[
-                    :, : self.oshape[3] * self.srf : self.srf
-                ]
-                
-                out[p_idx, slit_idx, :, :] = self.instr.pce[
-                    ..., np.newaxis
-                ]* self.wblur(sliced).sum(axis=2)
-        print("Return")   
-        return
     
-    def forward_parallel_static(inarray_f: array, idx: int) -> array:
+    def forward(self, inarray_f):
         """inarray is supposed in global coordinate, spatially blurred and in Fourier space.
 
         Output is an array of shape (pointing, slit, wavelength, alpha)."""
         # [pointing, slit, λ', α]
-        out = np.empty(self.oshape)
-        print("Idx = ", idx)
+        out = shared_dict.attach(self._metadata_path)["fw_data"]
         blurred = self.sblur(inarray_f[self.wslice, ...])
-        #logger.info(f"{self.name} : IDFT2({inarray_f.shape})")
         for p_idx, pointing in enumerate(self.pointings):
-            #logger.info(
-            #    f"{self.name} : gridding [{p_idx}/{len(self.pointings)}] {blurred.shape} -> {(blurred.shape[0],) + self.local_shape[1:]}"
-            #)
             gridded = self.gridding(blurred, pointing)
-            # Slit parallelization
             for slit_idx in range(self.instr.n_slit):
                 # Slicing, weighting and α subsampling for SR
-                #logger.info(f"{self.name} : slicing [{slit_idx+1}/{self.instr.n_slit}]")
                 sliced = self.slicing(gridded, slit_idx)[
                     :, : self.oshape[3] * self.srf : self.srf
                 ]
@@ -606,66 +556,16 @@ class Channel(LinOp):
                 out[p_idx, slit_idx, :, :] = self.instr.pce[
                     ..., np.newaxis
                 ]* self.wblur(sliced).sum(axis=2)
-        print("Return")   
-        return
-                
 
-    def forward_cf_parallel(self, inarray_f: array, idx: int):
-        """inarray is supposed in global coordinate, spatially blurred and in Fourier space.
 
-        Output is an array of shape (pointing, slit, wavelength, alpha)."""
-        # [pointing, slit, λ', α]
-        out = np.empty(self.oshape)
-        #logger.info(f"{self.name} : IDFT2({inarray_f.shape})")
-        blurred = self.sblur(inarray_f[self.wslice, ...])
-        for p_idx, pointing in enumerate(self.pointings):            
-            CFAPP.runJob("gridding_id:%d.%d" %(idx, p_idx), self._cf_forward_worker, args=(idx, p_idx, blurred, pointing))
-                
-
-    def _cf_forward_worker(self, idx, p_idx, blurred, pointing):
-           
-        out = np.empty(self.oshape[1:])
-        gridded = self.gridding(blurred, pointing)
-        for slit_idx in range(self.instr.n_slit):
-            sliced = self.slicing(gridded, slit_idx)[
-                :, : self.oshape[3] * self.srf : self.srf
-            ]
-            # λ blurring and Σ_β
-            out[slit_idx, :, :] = self.instr.pce[
-                    ..., np.newaxis
-                ] * self.wblur(sliced).sum(axis=2)
-
-        return {"idx": idx, "p_idx": p_idx, "out" : out}
-
-    def _forward_worker(self, idx, p_idx, blurred, pointing):
-        
-        out = np.empty(self.oshape[1:])
-        gridded = self.gridding(blurred, pointing)
-        for slit_idx in range(self.instr.n_slit):
-            sliced = self.slicing(gridded, slit_idx)[
-                :, : self.oshape[3] * self.srf : self.srf
-            ]
-            # λ blurring and Σ_β
-            out[slit_idx, :, :] = self.instr.pce[
-                    ..., np.newaxis
-                ] * self.wblur(sliced).sum(axis=2)
-
-        return {"idx": idx, "p_idx": p_idx, "out" : out.ravel()}
-     
-
-    def adjoint(self, measures: array) -> array:
-        out = np.zeros(self.ishape, dtype=np.complex128)
+    def adjoint(self, measures):
+        out = shared_dict.attach(self._metadata_path)["ad_data"]
         blurred = np.zeros(self.cshape)
         for p_idx, pointing in enumerate(self.pointings):
-            #logger.info(f"{self.name} : pointing [{p_idx+1}/{len(self.pointings)}]")
             gridded = np.zeros(self.local_shape)
             for slit_idx in range(self.instr.n_slit):
-                #logger.info(f"{self.name} : slicing [{slit_idx+1}/{self.instr.n_slit}]")
                 sliced = np.zeros(self.slit_shape(slit_idx))
                 # α zero-filling, λ blurrling_t, and β duplication
-                #logger.info(
-                #    f"{self.name} : wblur^T {measures.shape[2:] + (sliced.shape[2],)}"
-                #)
                 tmp = np.repeat(
                         np.expand_dims(
                             measures[p_idx, slit_idx] * self.instr.pce[..., np.newaxis],
@@ -679,54 +579,11 @@ class Channel(LinOp):
 
                 sliced[:, : self.oshape[3] * self.srf : self.srf] = tmp2
                     
-
                 gridded += self.slicing_t(sliced, slit_idx)
-            #logger.info(
-            #    f"{self.name} : gridding^T [{p_idx+1}/{len(self.pointings)}] {gridded.shape} -> {blurred.shape}"
-            #)
             blurred += self.gridding_t(gridded, pointing)
-        #logger.info(f"{self.name} : DFT2({blurred.shape})")
         out[self.wslice, ...] = self.sblur_t(blurred)
-        return out
 
-    def adjoint_multiproc(self, measures, idx, q):
-        print("Idx = ", idx)
-        out = np.zeros(self.ishape, dtype=np.complex128)
-        blurred = np.zeros(self.cshape)
-        for p_idx, pointing in enumerate(self.pointings):
-            #logger.info(f"{self.name} : pointing [{p_idx+1}/{len(self.pointings)}]")
-            gridded = np.zeros(self.local_shape)
-            for slit_idx in range(self.instr.n_slit):
-                #logger.info(f"{self.name} : slicing [{slit_idx+1}/{self.instr.n_slit}]")
-                sliced = np.zeros(self.slit_shape(slit_idx))
-                # α zero-filling, λ blurrling_t, and β duplication
-                #logger.info(
-                #    f"{self.name} : wblur^T {measures.shape[2:] + (sliced.shape[2],)}"
-                #)
-                tmp = np.repeat(
-                        np.expand_dims(
-                            measures[p_idx, slit_idx] * self.instr.pce[..., np.newaxis],
-                            axis=2,
-                        ),
-                        sliced.shape[2],
-                        axis=2,
-                    )
-
-                tmp2 = self.wblur_t(tmp)
-
-                sliced[:, : self.oshape[3] * self.srf : self.srf] = tmp2
-                    
-
-                gridded += self.slicing_t(sliced, slit_idx)
-            #logger.info(
-            #    f"{self.name} : gridding^T [{p_idx+1}/{len(self.pointings)}] {gridded.shape} -> {blurred.shape}"
-            #)
-            blurred += self.gridding_t(gridded, pointing)
-        #logger.info(f"{self.name} : DFT2({blurred.shape})")
-        #out[self.wslice, ...] = self.sblur_t(blurred)
-        return
-
-class Spectro(LinOp):
+class Spectro_shared(LinOp):
     def __init__(
         self,
         instrs: List[instru.IFU],
@@ -751,6 +608,12 @@ class Spectro(LinOp):
             self.step,
         )
 
+        _shared_metadata = shared_dict.create("s_metadata")
+        self._shared_metadata = _shared_metadata
+        for instr in instrs:
+            print("Create Chan shared dict named ", instr.get_name_pix())
+            _shared_metadata.addSubdict(instr.get_name_pix())
+
         self.channels = [
             Channel(
                 instr,
@@ -759,32 +622,15 @@ class Spectro(LinOp):
                 wavel_axis,
                 srf,
                 pointings,
+                _shared_metadata[instr.get_name_pix()].path,
             )
             for srf, instr in zip(srfs, instrs)
         ]
-        GD["channels"] = {chan.instr.name : {} for chan in self.channels}
-        for iChan, chan in enumerate(self.channels):
-            GD["channels"] = {iChan : {}}
-            GD["channels"][iChan]["name"] = chan.instr.name
-            GD["channels"][iChan]["oshape"] = chan.oshape
-            GD["channels"][iChan]["ishape"] = chan.ishape
-            GD["channels"][iChan]["srf"] = chan.srf
-
-
-        """ _data = shared_dict.create("data")
-        self._data_path = _data.path
-        print("DATA PATH IS ", self._data_path)
-        for iChan, chan in enumerate(self.channels):
-            _data.addSubdict(iChan)
-            _data[iChan]["data"] = np.zeros(chan.oshape) """
 
         self._idx = np.cumsum([0] + [np.prod(chan.oshape) for chan in self.channels])
         self.imshape = (len(alpha_axis), len(beta_axis))
         ishape = (len(wavel_axis), len(alpha_axis), len(beta_axis))
         oshape = (self._idx[-1],)
-
-        # Build Genereal dictionnary filled with all metadata
-        self.GD = GD
 
         super().__init__(ishape, oshape, "Spectro")
 
@@ -801,109 +647,43 @@ class Spectro(LinOp):
     def set_general_dic(self):
 
         return
+  
 
     def forward(self, inarray: array) -> array:
         out = np.zeros(self.oshape)
-        #logger.info(f"Spatial blurring DFT2({inarray.shape})")
+        logger.info(f"Spatial blurring DFT2({inarray.shape})")
         blurred_f = dft(inarray) * self.sotf
         for idx, chan in enumerate(self.channels):
             logger.info(f"Channel {chan.name}")
-            out[self._idx[idx] : self._idx[idx + 1]] = chan.forward(blurred_f).ravel()
+            APPL.runJob("forward_id:%d"%idx, chan.forward_multiproc, args=(blurred_f,), serial=False)
 
-        return out
-    
-    def forward_multiproc(self, inarray: array) -> array:
-        out = np.zeros(self.oshape)
-        #logger.info(f"Spatial blurring DFT2({inarray.shape})")
-        blurred_f = dft(inarray) * self.sotf
-        results = []
-        q = Queue()
-        p = []
-        for idx, chan in enumerate(self.channels):
-            logger.info(f"Channel {chan.name}")
-            p.append(Process(target=chan.forward_multiproc, args=(blurred_f, idx, q)))
-
-        for proc in p:
-            proc.start()
-
-        for j in p:
-            j.join()
+        APPL.awaitJobResult()
         
-
-        #connection.wait(proc.sentinel for proc in p)
-        #for idx, _ in enumerate(self.channels):
-
-            #out[self._idx[idx] : self._idx[idx + 1]] = chan.forward(blurred_f).ravel()
-
-        """ for dicoresults in res:
-            out[self._idx[dicoresults["idx"]] : self._idx[dicoresults["idx"] +1]] = dicoresults["out"].ravel()
-
-        return out """
-
-
-    def forward_parallel(self, inarray: array) -> array:
-        out = np.zeros(self.oshape)
-        logger.info(f"Spatial blurring DFT2({inarray.shape})")
-        blurred_f = dft(inarray) * self.sotf
+        self._shared_metadata.reload()
         for idx, chan in enumerate(self.channels):
-            logger.info(f"Channel {chan.name}")
-            APP.runJob("gridding_id:%d" %(idx), chan.forward_parallel, args=(blurred_f, idx))
+            fw_data = self._shared_metadata[chan.name]["fw_data"]
+            out[self._idx[idx] : self._idx[idx + 1]] = fw_data.ravel()
 
-        APP.awaitJobResults("gridding_id:*")
-        #for dicoresults in results:
-        #    out[self._idx[dicoresults["idx"]] + dicoresults["p_idx"]*np.prod(dicoresults["out"].shape): self._idx[dicoresults["idx"]]+ (dicoresults["p_idx"]+1)*np.prod(dicoresults["out"].shape)] = dicoresults["out"].ravel()
+        return out 
 
-        return out
 
-    def forward_cf_parallel(self, inarray: array) -> array:
-        out = np.zeros(self.oshape)
-        logger.info(f"Spatial blurring DFT2({inarray.shape})")
-        blurred_f = dft(inarray) * self.sotf
-        for idx, chan in enumerate(self.channels):
-            logger.info(f"Channel {chan.name}")
-            chan.forward_cf_parallel(blurred_f, idx)
-
-        results = CFAPP.awaitJobResults("gridding_id:*")
-        for dicoresults in results:
-            out[self._idx[dicoresults["idx"]] + dicoresults["p_idx"]*np.prod(dicoresults["out"].shape): self._idx[dicoresults["idx"]]+ (dicoresults["p_idx"]+1)*np.prod(dicoresults["out"].shape)] = dicoresults["out"].ravel()
-
-        return out
-
-    
     def adjoint(self, inarray: array) -> array:
         tmp = np.zeros(
             self.ishape[:2] + (self.ishape[2] // 2 + 1,), dtype=np.complex128
         )
         for idx, chan in enumerate(self.channels):
             logger.info(f"Channel {chan.name}")
-            tmp += chan.adjoint(
-                np.reshape(inarray[self._idx[idx] : self._idx[idx + 1]], chan.oshape)
-            )
+            APPL.runJob("adjoint_id:%d"%idx, chan.adjoint_multiproc, args=(np.reshape(inarray[self._idx[idx] : self._idx[idx + 1]], chan.oshape),), serial=False)
+
+
+        APPL.awaitJobResult()
+        self._shared_metadata.reload()
+        for idx, chan in enumerate(self.channels):
+            ad_data = self._shared_metadata[chan.name]["ad_data"]
+            tmp += ad_data
+        
         logger.info(f"Spatial blurring^T : IDFT2({tmp.shape})")
         return idft(tmp * self.sotf.conj(), self.imshape)
-
-
-    def adjoint_multiproc(self, inarray: array) -> array:
-        tmp = np.zeros(
-            self.ishape[:2] + (self.ishape[2] // 2 + 1,), dtype=np.complex128
-        )
-
-        results = []
-        q = Queue()
-        p = []
-        for idx, chan in enumerate(self.channels):
-            logger.info(f"Channel {chan.name}")
-            p.append(Process(target=chan.adjoint_multiproc, args=(np.reshape(inarray[self._idx[idx] : self._idx[idx + 1]], chan.oshape), idx, q)))
-
-        for proc in p:
-            proc.start()
-
-        for j in p:
-            j.join()
-
-        #logger.info(f"Spatial blurring^T : IDFT2({tmp.shape})")
-        #return idft(tmp * self.sotf.conj(), self.imshape)
-
 
 
     def qdcoadd(self, measures: array) -> array:
